refactor out helper methods in reports
